+++ server/src/main/java/org/zanata/magpie/api/security/CheckRoleDecisionVoter.java
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA, or see the FSF
  * site: http://www.fsf.org.
  */
-package org.zanata.magpie.security;
+package org.zanata.magpie.api.security;
 
 import java.util.Optional;
 import java.util.Set;
