+++ server/src/main/java/org/zanata/magpie/api/service/impl/AccountResourceImpl.java
@@ -46,7 +46,7 @@ import org.zanata.magpie.api.dto.AccountDto;
 import org.zanata.magpie.api.dto.CredentialDto;
 import org.zanata.magpie.exception.DataConstraintViolationException;
 import org.zanata.magpie.exception.MTException;
-import org.zanata.magpie.security.AccountCreated;
+import org.zanata.magpie.event.AccountCreated;
 import org.zanata.magpie.service.AccountService;
 
 import com.google.common.base.Throwables;
