+++ server/src/main/java/org/zanata/magpie/event/AccountCreated.java
@@ -18,7 +18,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.zanata.magpie.security;
+package org.zanata.magpie.event;
 
 import java.util.Objects;
 import java.util.Set;
