+++ server/src/main/java/org/zanata/magpie/service/MTStartup.java
@@ -29,7 +29,7 @@ import org.zanata.magpie.api.dto.AccountDto;
 import org.zanata.magpie.exception.MTException;
 import org.zanata.magpie.model.BackendID;
 import org.zanata.magpie.model.Role;
-import org.zanata.magpie.security.AccountCreated;
+import org.zanata.magpie.event.AccountCreated;
 import org.zanata.magpie.util.PasswordUtil;
 
 import static org.zanata.magpie.service.ResourceProducer.REPLICATE_CACHE;
