+++ server/src/test/java/org/zanata/magpie/api/service/impl/AccountResourceImplTest.java
@@ -26,9 +26,9 @@ import org.zanata.magpie.api.dto.CredentialDto;
 import org.zanata.magpie.exception.DataConstraintViolationException;
 import org.zanata.magpie.exception.MTException;
 import org.zanata.magpie.model.AccountType;
-import org.zanata.magpie.security.AccountCreated;
+import org.zanata.magpie.event.AccountCreated;
 import org.zanata.magpie.service.AccountService;
-import org.zanata.magpie.util.ValidatorProducer;
+import org.zanata.magpie.service.ResourceProducer;
 
 
 public class AccountResourceImplTest {
@@ -41,7 +41,7 @@ public class AccountResourceImplTest {
     public void setUp() {
         MockitoAnnotations.initMocks(this);
         accountResource = new AccountResourceImpl(accountService,
-                new ValidatorProducer().getValidator(), accountCreatedEvent);
+                new ResourceProducer().getValidator(), accountCreatedEvent);
         accountResource.uriInfo = new ResteasyUriInfo("account", "", "");
     }
